We allocate 4kb of space each time and this space won't be collected by GC until we remove references to syscall.NetlinkMessages, because it stores parts of original slice. Also allocation of 4kb is unneeded every time and in common cases makes call slower.
This patch though makes receiving big netlink messages slightly slower(because of copying answer to new slice), but allows to save significant amount of memory.
Here results of comparing benchmarks:
```
benchmark                 old ns/op     new ns/op     delta
BenchmarkLinkAddDel-4     6860622       7510755       +9.48%
BenchmarkLinkList-4       59623         52799         -11.45%

benchmark                 old allocs     new allocs     delta
BenchmarkLinkAddDel-4     138            144            +4.35%
BenchmarkLinkList-4       39             41             +5.13%

benchmark                 old bytes     new bytes     delta
BenchmarkLinkAddDel-4     34632         14237         -58.89%
BenchmarkLinkList-4       13824         8301          -39.95%
```