@LK4D4 2015-06-17
I did some measurements with docker and it seems that we have 70-99% of thrash memory on each call, which is pretty bad in Go, because memory released to OS not immediately. What if I introduce threshold(like 1.5k) for using slice from pool, so in worst case we will have 60% of thrash memory with decent performance(I think it should be pretty equal to old performance).