@vishvananda 2015-06-19
I'm starting to think the optimization isn't really worth it if we are just saving a few bytes of memory usage.