@LK4D4 2015-06-19
@vishvananda It's almost always around 4kb per netlink call. But maybe you're right, it's too slippery way.